requirements.txt
alembic==0.9.7
attribdict==1.0.0
alphai_watson
amqp==2.2.2
appnope==0.1.0
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
click==6.7
decorator==4.2.1
Flask==0.12.2
Flask-Celery-Helper==1.1.0
gunicorn
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
kombu==4.1.0
Mako==1.0.7
MarkupSafe==1.0
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
python-dotenv==0.7.1
python-editor==1.0.3
pytz==2017.3
redis==2.10.6
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.2
sqlalchemy-pagination
traitlets==4.3.2
vine==1.1.4
wcwidth==0.1.7
Werkzeug==0.14.1
marshmallow==3.0.0b4
marshmallow-enum==1.4.1
requests
passlib==1.7.1
pandas==0.22
tables==3.4.2
numpy
sklearn
pyyaml
alphai_model_store==0.0.4
alphai-es-datasource==0.0.4
